Migrated to features and build fixes
