+++ packages/ui/v2/core/ImageUploader.tsx
@@ -1,14 +1,61 @@
+import * as SliderPrimitive from "@radix-ui/react-slider";
 import { FormEvent, useCallback, useEffect, useState } from "react";
 import Cropper from "react-easy-crop";
 
+import { useLocale } from "@calcom/lib/hooks/useLocale";
 import { Dialog, DialogClose, DialogContent, DialogTrigger } from "@calcom/ui/Dialog";
 import Button from "@calcom/ui/v2/core/Button";
 
-import { Area, getCroppedImg } from "@lib/cropImage";
-import { useFileReader } from "@lib/hooks/useFileReader";
-import { useLocale } from "@lib/hooks/useLocale";
+type ReadAsMethod = "readAsText" | "readAsDataURL" | "readAsArrayBuffer" | "readAsBinaryString";
 
-import Slider from "@components/Slider";
+type UseFileReaderProps = {
+  method: ReadAsMethod;
+  onLoad?: (result: unknown) => void;
+};
+
+type Area = {
+  width: number;
+  height: number;
+  x: number;
+  y: number;
+};
+
+const MAX_IMAGE_SIZE = 512;
+
+const useFileReader = (options: UseFileReaderProps) => {
+  const { method = "readAsText", onLoad } = options;
+  const [file, setFile] = useState<File | null>(null);
+  const [loading, setLoading] = useState<boolean>(false);
+  const [error, setError] = useState<DOMException | null>(null);
+  const [result, setResult] = useState<string | ArrayBuffer | null>(null);
+
+  useEffect(() => {
+    if (!file && result) {
+      setResult(null);
+    }
+  }, [file, result]);
+
+  useEffect(() => {
+    if (!file) {
+      return;
+    }
+
+    const reader = new FileReader();
+    reader.onloadstart = () => setLoading(true);
+    reader.onloadend = () => setLoading(false);
+    reader.onerror = () => setError(reader.error);
+
+    reader.onload = (e: ProgressEvent<FileReader>) => {
+      setResult(e.target?.result ?? null);
+      if (onLoad) {
+        onLoad(e.target?.result ?? null);
+      }
+    };
+    reader[method](file);
+  }, [file, method, onLoad]);
+
+  return [{ result, error, file, loading }, setFile] as const;
+};
 
 type ImageUploaderProps = {
   id: string;
@@ -165,3 +212,75 @@ export default function ImageUploader({
     </Dialog>
   );
 }
+
+const createImage = (url: string) =>
+  new Promise<HTMLImageElement>((resolve, reject) => {
+    const image = new Image();
+    image.addEventListener("load", () => resolve(image));
+    image.addEventListener("error", (error) => reject(error));
+    image.setAttribute("crossOrigin", "anonymous"); // needed to avoid cross-origin issues on CodeSandbox
+    image.src = url;
+  });
+
+async function getCroppedImg(imageSrc: string, pixelCrop: Area): Promise<string> {
+  const image = await createImage(imageSrc);
+  const canvas = document.createElement("canvas");
+  const ctx = canvas.getContext("2d");
+  if (!ctx) throw new Error("Context is null, this should never happen.");
+
+  const maxSize = Math.max(image.naturalWidth, image.naturalHeight);
+  const resizeRatio = MAX_IMAGE_SIZE / maxSize < 1 ? Math.max(MAX_IMAGE_SIZE / maxSize, 0.75) : 1;
+  // huh, what? - Having this turned off actually improves image quality as otherwise anti-aliasing is applied
+  // this reduces the quality of the image overall because it anti-aliases the existing, copied image; blur results
+  ctx.imageSmoothingEnabled = false;
+  // pixelCrop is always 1:1 - width = height
+  canvas.width = canvas.height = Math.min(maxSize * resizeRatio, pixelCrop.width);
+
+  ctx.drawImage(
+    image,
+    pixelCrop.x,
+    pixelCrop.y,
+    pixelCrop.width,
+    pixelCrop.height,
+    0,
+    0,
+    canvas.width,
+    canvas.height
+  );
+
+  // on very low ratios, the quality of the resize becomes awful. For this reason the resizeRatio is limited to 0.75
+  if (resizeRatio <= 0.75) {
+    // With a smaller image, thus improved ratio. Keep doing this until the resizeRatio > 0.75.
+    return getCroppedImg(canvas.toDataURL("image/png"), {
+      width: canvas.width,
+      height: canvas.height,
+      x: 0,
+      y: 0,
+    });
+  }
+
+  return canvas.toDataURL("image/png");
+}
+
+const Slider = ({
+  value,
+  label,
+  changeHandler,
+  ...props
+}: Omit<SliderPrimitive.SliderProps, "value"> & {
+  value: number;
+  label: string;
+  changeHandler: (value: number) => void;
+}) => (
+  <SliderPrimitive.Root
+    className="slider mt-2"
+    value={[value]}
+    aria-label={label}
+    onValueChange={(value: number[]) => changeHandler(value[0] ?? value)}
+    {...props}>
+    <SliderPrimitive.Track className="slider-track">
+      <SliderPrimitive.Range className="slider-range" />
+    </SliderPrimitive.Track>
+    <SliderPrimitive.Thumb className="slider-thumb" />
+  </SliderPrimitive.Root>
+);
